work on Auth view
