+++ src/App/components/Reuse/Button.tsx
@@ -1,10 +1,13 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
 import React from 'react';
 import classes from './Button.module.css';
 
 function Button(props) {
   return (
-    <button className={classes.button} type="submit">
-      {/* eslint-disable-next-line react/destructuring-assignment,react/prop-types */}
+    // eslint-disable-next-line
+    <button className={classes.button} type={props.type}>
+      {/* eslint-disable-next-line */}
       {props.text}
     </button>
   );
