+++ src/App/components/User/UserProfile.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function UserProfile() {
+  return (
+    <main>
+      <h2>My User Profile</h2>
+    </main>
+  );
+}
+
+export default UserProfile;
