+++ src/index.css
@@ -1,3 +1,4 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@400;700&display=swap');
 
 * {
     box-sizing: border-box;
@@ -11,3 +12,18 @@ body {
     margin: 0;
     background-color: white;
 }
+
+main {
+    margin-top: 7rem;
+    text-align: center;
+}
+
+h1,
+h2,
+h3,
+p {
+    color: #042b5f;
+}
+
+
+
