+++ src/index.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-// eslint-disable-next-line import/no-unresolved
 import ReactDOM from 'react-dom/client';
+
 import './index.css';
-// eslint-disable-next-line import/extensions
 import App from './App/App';
-// eslint-disable-next-line import/extensions
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
